feat: transfer avatar component functions to ts file (unfinished)
